moved auto redirection to dashboard from landing page to demo page
